requirements.txt
flask==3.0.3
pydantic==2.8.2
email-validator==2.2.0
python-dotenv==1.0.1
pypdf==4.3.1
numpy==1.26.4
pandas==2.2.2
langchain==0.2.16
langchain-openai==0.1.23
langgraph==0.2.28
Flask-Cors>=4.0
openai>=1.0.0
pymupdf>=1.23.0
tqdm>=4.66.0
chromadb>=0.5.0

